add an option to add new lines at the top when the TTY is vertically expanded (#167)
To enable:

    zstyle ':z4h:' term-vresize 'top'

Open a new terminal tab, run `seq 1000` and resize the terminal to make it
taller. Prompt should stay at the bottom and a part of scrollback that was
hidden prior to resizing should now be visible. In other words, additional
lines should be added to the terminal at the top.

The default value is 'bottom'. In this case new empty lines get added at the
bottom of the terminal so that prompt remains on the same line number.

In the future, if the option isn't set explicitly, zsh4humans may try to
guess which value matches the underlying terminal's behavior.